requirements.txt
Django==4.1.2
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-filter==22.1
django-money==3.0.0
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.4
